refactor: 카드 타입을 체크하는 함수 분리
